+++ PhpSerializerNET/PhpSerialization.cs
@@ -9,6 +9,7 @@
 
 using System;
 using System.Collections.Generic;
+using System.Text;
 
 namespace PhpSerializerNET;
 
@@ -62,6 +63,27 @@ public static class PhpSerialization {
 		return new PhpDeserializer(tokens, inputBytes, options).Deserialize();
 	}
 
+	public static object? DeserializeUtf8(
+		ReadOnlySpan<byte> input,
+		PhpDeserializationOptions? options = null
+	) {
+		if (input.Length == 0) {
+			throw new ArgumentOutOfRangeException
+				(nameof(input),
+				"PhpSerialization.DeserializeUtf8(): Parameter 'input' must not be empty."
+			);
+		}
+		if (options == null)  {
+			options = PhpDeserializationOptions.DefaultOptions;
+		} else if (options.InputEncoding != Encoding.UTF8) {
+			throw new ArgumentException("Can not use input encoding other than UTF8", nameof(options));
+		}
+		int tokenCount = PhpTokenValidator.Validate(input);
+		Span<PhpToken> tokens = new PhpToken[tokenCount];
+		PhpTokenizer.Tokenize(input, tokens);
+		return new PhpDeserializer(tokens, input, options).Deserialize();
+	}
+
 	/// <summary>
 	/// The serialized data to deserialize.
 	/// </summary>
@@ -78,26 +100,12 @@ public static class PhpSerialization {
 	/// <returns>
 	/// The deserialized object.
 	/// </returns>
-	public static T Deserialize<T>(
-		string input,
-		PhpDeserializationOptions? options = null
-	) {
-		if (string.IsNullOrEmpty(input)) {
-			throw new ArgumentOutOfRangeException(nameof(input), "PhpSerialization.Deserialize(): Parameter 'input' must not be null or empty.");
-		}
-		if (options == null)  {
-			options = PhpDeserializationOptions.DefaultOptions;
-		}
-		int size = options.InputEncoding.GetByteCount(input);
-		Span<byte> inputBytes = size < 256
-			? stackalloc byte[size]
-			: new byte[size];
-		options.InputEncoding.GetBytes(input, inputBytes);
-		int tokenCount = PhpTokenValidator.Validate(inputBytes);
-		Span<PhpToken> tokens = new PhpToken[tokenCount];
-		PhpTokenizer.Tokenize(inputBytes, tokens);
-		return new PhpDeserializer(tokens, inputBytes, options).Deserialize<T>();
-	}
+	public static T? Deserialize<T>(string input, PhpDeserializationOptions? options = null)
+		=> (T?)Deserialize(input, typeof(T), options);
+
+	public static T? DeserializeUtf8<T>(ReadOnlySpan<byte> input, PhpDeserializationOptions? options = null)
+		=> (T?)DeserializeUtf8(input, typeof(T), options);
+
 
 	/// <summary>
 	/// The serialized data to deserialize.
@@ -120,9 +128,6 @@ public static class PhpSerialization {
 		Type type,
 		PhpDeserializationOptions? options = null
 	) {
-		if (type == null) {
-			throw new ArgumentNullException(nameof(type));
-		}
 		if (string.IsNullOrEmpty(input)) {
 			throw new ArgumentOutOfRangeException(nameof(input), "PhpSerialization.Deserialize(): Parameter 'input' must not be null or empty.");
 		}
@@ -138,6 +143,44 @@ public static class PhpSerialization {
 		Span<PhpToken> tokens = new PhpToken[tokenCount];
 		PhpTokenizer.Tokenize(inputBytes, tokens);
 		return new PhpDeserializer(tokens, inputBytes, options).Deserialize(type);
+	}
+
+		/// <summary>
+	/// The serialized data to deserialize.
+	/// </summary>
+	/// <param name="input">
+	/// Data in the PHP de/serialization format.
+	/// </param>
+	/// <param name="options">
+	/// Options for deserialization. See the  <see cref="PhpDeserializationOptions"/> class for more details.
+	/// </param>
+	/// <param name="type">
+	/// The desired output type.
+	/// This should be one of the primitives or a class with a public parameterless constructor.
+	/// </typeparam>
+	/// <returns>
+	/// The deserialized object.
+	/// </returns>
+	public static object? DeserializeUtf8(
+		ReadOnlySpan<byte> input,
+		Type type,
+		PhpDeserializationOptions? options = null
+	) {
+		if (input.Length == 0) {
+			throw new ArgumentOutOfRangeException
+				(nameof(input),
+				"PhpSerialization.DeserializeUtf8(): Parameter 'input' must not be empty."
+			);
+		}
+		if (options == null)  {
+			options = PhpDeserializationOptions.DefaultOptions;
+		} else if (options.InputEncoding != Encoding.UTF8) {
+			throw new ArgumentException("Can not use input encoding other than UTF8", nameof(options));
+		}
+		int tokenCount = PhpTokenValidator.Validate(input);
+		Span<PhpToken> tokens = new PhpToken[tokenCount];
+		PhpTokenizer.Tokenize(input, tokens);
+		return new PhpDeserializer(tokens, input, options).Deserialize(type);
 	}
 
 	/// <summary>